Change the checkpoint id to an unsigned long long

+++ src/hs_checkpoint_reader.h
@@ -56,18 +56,18 @@ void hs_lookup_input_checkpoint(hs_checkpoint_reader* cpr,
                                 const char* key,
                                 const char* path,
                                 const char* subdir,
-                                size_t* id,
+                                unsigned long long* id,
                                 size_t* offset);
 
 void hs_update_input_checkpoint(hs_checkpoint_reader* cpr,
                                 const char* key,
                                 const char* subdir,
-                                size_t id,
+                                unsigned long long id,
                                 size_t offset);
 
 void hs_update_id_checkpoint(hs_checkpoint_reader* cpr,
                              const char* key,
-                             size_t id);
+                             unsigned long long id);
 
 void hs_output_checkpoints(hs_checkpoint_reader* cpr, FILE* fh);
